# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TechSpar",
    "version": "0.4.1"
  },
  "components": {
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "title": "ValidationError",
        "properties": {
          "loc": {
            "type": "array",
            "title": "Location",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ]
      },
      "HTTPValidationError": {
        "type": "object",
        "title": "HTTPValidationError",
        "properties": {
          "detail": {
            "type": "array",
            "title": "Detail",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/api/copilot/prep": {
      "post": {
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "jd_text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "company": {
                    "type": "string",
                    "default": ""
                  },
                  "position": {
                    "type": "string",
                    "default": ""
                  },
                  "document_ids": {
                    "type": "string",
                    "default": "[]"
                  }
                },
                "required": [
                  "jd_text"
                ]
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "jd_text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "company": {
                    "type": "string",
                    "default": ""
                  },
                  "position": {
                    "type": "string",
                    "default": ""
                  },
                  "document_ids": {
                    "type": "string",
                    "default": "[]"
                  }
                },
                "required": [
                  "jd_text"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Queued Copilot preparation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prep_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "prep_id"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/preps": {
      "get": {
        "responses": {
          "200": {
            "description": "Copilot preparations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/prep/{prep_id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Copilot preparation status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "nullable": true
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted Copilot preparation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/prep/{prep_id}/tree": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Copilot strategy tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "nullable": true
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/prep/{prep_id}/prepared-answers": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Compiled interview knowledge package",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    },
                    "document_ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "source_snapshot": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "document_id": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "document_id",
                          "filename",
                          "updated_at"
                        ]
                      }
                    },
                    "source_fingerprint": {
                      "type": "string"
                    },
                    "prepared_answers": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "answer_id": {
                            "type": "string"
                          },
                          "node_id": {
                            "type": "string"
                          },
                          "topic": {
                            "type": "string"
                          },
                          "intent": {
                            "type": "string"
                          },
                          "question_variants": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "prepared_answer": {
                            "type": "string"
                          },
                          "short_answer": {
                            "type": "string"
                          },
                          "key_points": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "source_refs": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "source_type": {
                                  "type": "string",
                                  "enum": [
                                    "personal_document",
                                    "resume",
                                    "profile"
                                  ]
                                },
                                "document_id": {
                                  "type": "string"
                                },
                                "filename": {
                                  "type": "string"
                                },
                                "evidence": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "source_type",
                                "evidence"
                              ]
                            }
                          },
                          "confidence": {
                            "type": "number",
                            "minimum": 0,
                            "maximum": 1
                          },
                          "usable": {
                            "type": "boolean"
                          },
                          "warnings": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        },
                        "required": [
                          "answer_id",
                          "node_id",
                          "topic",
                          "intent",
                          "question_variants",
                          "prepared_answer",
                          "short_answer",
                          "key_points",
                          "source_refs",
                          "confidence",
                          "usable",
                          "warnings"
                        ]
                      }
                    },
                    "uncompiled_nodes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "node_id": {
                            "type": "string"
                          },
                          "error": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "node_id",
                          "error"
                        ]
                      }
                    },
                    "compile_stats": {
                      "type": "object",
                      "properties": {
                        "strategy_nodes": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "compiled_answers": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "question_variants": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "source_documents": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "strategy_nodes",
                        "compiled_answers",
                        "question_variants",
                        "source_documents"
                      ]
                    },
                    "index_status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "ready",
                        "error"
                      ]
                    },
                    "index_error": {
                      "type": "string"
                    },
                    "source_state": {
                      "type": "string",
                      "enum": [
                        "current",
                        "stale"
                      ]
                    }
                  },
                  "required": [
                    "version",
                    "document_ids",
                    "source_snapshot",
                    "source_fingerprint",
                    "prepared_answers",
                    "uncompiled_nodes",
                    "compile_stats",
                    "index_status",
                    "source_state"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/prep/{prep_id}/test-match": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "question"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Prepared answer matcher result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "matched": {
                      "type": "boolean"
                    },
                    "route": {
                      "type": "string",
                      "enum": [
                        "prepared",
                        "partial",
                        "miss"
                      ]
                    },
                    "score": {
                      "type": "number"
                    },
                    "semantic_score": {
                      "type": "number"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "node_id": {
                      "type": "string"
                    },
                    "answer_id": {
                      "type": "string"
                    },
                    "matched_question": {
                      "type": "string"
                    },
                    "answer": {
                      "type": "object",
                      "properties": {
                        "answer_id": {
                          "type": "string"
                        },
                        "node_id": {
                          "type": "string"
                        },
                        "topic": {
                          "type": "string"
                        },
                        "intent": {
                          "type": "string"
                        },
                        "question_variants": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "prepared_answer": {
                          "type": "string"
                        },
                        "short_answer": {
                          "type": "string"
                        },
                        "key_points": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "source_refs": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "source_type": {
                                "type": "string",
                                "enum": [
                                  "personal_document",
                                  "resume",
                                  "profile"
                                ]
                              },
                              "document_id": {
                                "type": "string"
                              },
                              "filename": {
                                "type": "string"
                              },
                              "evidence": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "source_type",
                              "evidence"
                            ]
                          }
                        },
                        "confidence": {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 1
                        },
                        "usable": {
                          "type": "boolean"
                        },
                        "warnings": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "answer_id",
                        "node_id",
                        "topic",
                        "intent",
                        "question_variants",
                        "prepared_answer",
                        "short_answer",
                        "key_points",
                        "source_refs",
                        "confidence",
                        "usable",
                        "warnings"
                      ]
                    },
                    "candidates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "score": {
                            "type": "number"
                          },
                          "semantic_score": {
                            "type": "number"
                          },
                          "lexical_score": {
                            "type": "number"
                          },
                          "node_id": {
                            "type": "string"
                          },
                          "answer_id": {
                            "type": "string"
                          },
                          "matched_question": {
                            "type": "string"
                          },
                          "compatible": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "score",
                          "semantic_score",
                          "lexical_score",
                          "node_id",
                          "answer_id",
                          "matched_question",
                          "compatible"
                        ]
                      }
                    },
                    "latency_ms": {
                      "type": "number",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "matched",
                    "route",
                    "score",
                    "candidates",
                    "latency_ms"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  }
}